Added ability to change timeout value for Groovy Scripting. Possibility to choose TLS version for SSL connections: v1, v1.1 and v1.2
* making default value visible as TLSv1 (first one from the list)
* default timeout set to 6 seconds

+++ src/com/intellij/idea/plugin/hybris/tools/remote/http/AbstractHybrisHacHttpClient.java
@@ -63,7 +63,7 @@ import static org.apache.http.HttpVersion.HTTP_1_1;
 public abstract class AbstractHybrisHacHttpClient {
 
     private static final Logger LOG = Logger.getInstance(AbstractHybrisHacHttpClient.class);
-    public static final long DEFAULT_HAC_TIMEOUT = 6000L;
+    public static final int DEFAULT_HAC_TIMEOUT = 6000;
     protected String sessionId;
 
     public String login(Project project) {
